requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
PyPDF2
